+++ Tweaks/UiAdjustment/TargetCastbarCountdown.cs
@@ -80,7 +80,7 @@ public unsafe class TargetCastbarCountdown : UiAdjustments.SubTweak
     
     public override void Setup()
     {
-        AddChangelogNewTweak(Changelog.UnreleasedVersion).Author("MidoriKami");
+        AddChangelogNewTweak("1.8.3.0");
         base.Setup();
     }